refactor: 테스트파일 패키지 이동
