+++ app/src/test/java/woowacourse/movie/model/movie/screening/DefaultShowtimePolicyTest.kt
@@ -1,7 +1,6 @@
-package woowacourse.movie.model
+package woowacourse.movie.model.movie.screening
 
 import org.assertj.core.api.Assertions.assertThat
-import woowacourse.movie.model.movie.screening.DefaultShowtimePolicy
 import java.time.LocalDate
 import java.time.LocalTime
 import kotlin.test.Test
