+++ app/src/test/java/woowacourse/movie/model/ticket/TicketCountTest.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.model
+package woowacourse.movie.model.movie.ticket
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
